requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF processing
pdfplumber>=0.11.0
PyPDF2>=3.0.0

# Data handling
pyyaml>=6.0
jsonschema>=4.19.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Database (optional)
sqlite-utils>=3.35

# CLI utilities
click>=8.1.0
tqdm>=4.66.0
colorama>=0.4.6

# Data validation
python-dateutil>=2.8.0
validators>=0.22.0

# SSL/Security
urllib3>=2.0.0
certifi>=2023.7.0

# Export formats
openpyxl>=3.1.0
pandas>=2.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
